refactor : SellService -> SellNowService
